+++ cp_mod_migration_utilz_test.go
@@ -1,11 +1,16 @@
 package main_test
 
 import (
+	"errors"
+	"io"
 	"os"
+	"testing"
 
 	. "github.com/onsi/gomega"
-
+	"github.com/stretchr/testify/mock"
 	v294 "github.tools.sap/framefrog/cp-mod-migrator/pkg/cproxy/api/v294"
+	"github.tools.sap/framefrog/cp-mod-migrator/pkg/extract"
+	"github.tools.sap/framefrog/cp-mod-migrator/pkg/mocks"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/yaml"
@@ -73,4 +78,48 @@ func cp(name, namespace string) v294.ConnectivityProxy {
 			},
 		},
 	}
+}
+
+func loadCMs(data *[]corev1.ConfigMap) error {
+	file, err := os.Open("../../hack/testdata/configmaps.yaml")
+	if err != nil {
+		return err
+	}
+
+	decoder := yaml.NewYAMLOrJSONDecoder(file, 2048)
+	var cm corev1.ConfigMap
+	for {
+		err := decoder.Decode(&cm)
+		if errors.Is(err, io.EOF) {
+			break
+		}
+
+		if err != nil {
+			return err
+		}
+
+		*data = append(*data, cm)
+	}
+	return nil
+}
+
+func newClient(t *testing.T, cms []corev1.ConfigMap) extract.Client {
+	var index int
+	runFn := func(args mock.Arguments) {
+		cm := args.Get(2).(*corev1.ConfigMap)
+		*cm = cms[index]
+		if index == len(cms)-1 {
+			index = 0
+			return
+		}
+		index++
+	}
+
+	client := mocks.NewClient(t)
+	client.On("Get", mock.Anything, mock.Anything, mock.Anything).
+		Run(runFn).
+		Return(nil).
+		Times(len(cms))
+
+	return client
 }
